pyproject.toml
[project]
name = "DoPARSE"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "josemarcosrf", email = "jose.marcos.rf@gmail.com"},
]
dependencies = [
    "uvicorn>=0.34.0",
    "fastapi>=0.115.11",
    "python-multipart>=0.0.20",
    "ray[serve]>=2.43.0",
    "invoke>=2.2.0",
    "click>=8.1.8",
    "Pillow>=10.1.0",
    "torch>=2.9.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "isort>=6.0.0",
    "ruff>=0.9.7",
    "pre-commit>=4.1.0",
]
macos = [
    "mlx-vlm>=0.0.7",
]
marker = [
    "marker-pdf>=1.10.1",
]
docling = [
    "docling>=2.61.1",
    "deepsearch-glm>=1.0.0",
    "tokenizers>=0.22.1",
]

[tool.pdm]
distribution = false

[tool.pdm.scripts.check-lint]
shell = """
    ruff format --check .
    isort --check .
    ruff check .
"""
help = "Run linting and formating checks."

[tool.pdm.scripts.reformat]
shell = """
    isort .
    ruff format .
"""
help = "Reformat the code using isort and ruff."

[tool.pdm.scripts]
doparse = "python -m src.doparse"

[tool.ruff]
line-length = 88
output-format = "grouped"


[tool.isort]
profile = "black"
line_length = 88
filter_files = true
force_alphabetical_sort_within_sections = true

